Rename StyleResolver to StylesheetFactory
Stylesheet factory is just a better name...
